clarify language modules
