+++ src/languages/logic/parser.rs
@@ -1,13 +1,10 @@
-use std::convert::TryFrom;
-
 use decorum::R32;
 use rustc_hash::FxHashMap;
-use serde::Deserialize;
 use smallvec::SmallVec;
 
 use super::tokenizer::{Token, TokenType, Metadata};
 use crate::item::{Resource, Skill, Shard, Teleporter};
-use crate::util::{Difficulty, Glitch, RefillType, NodeType, Enemy, Position, UberState};
+use crate::util::{Difficulty, Glitch, RefillType, NodeType, Enemy, Position};
 
 #[derive(Debug)]
 pub struct ParseError {
@@ -541,81 +538,4 @@ pub fn parse_areas<'a>(tokens: Vec<Token<'a>>, metadata: &Metadata) -> Result<Ar
         regions,
         anchors,
     })
-}
-
-#[derive(Debug)]
-pub struct Location {
-    pub name: String,
-    pub zone: String,
-    pub uber_state: UberState,
-    pub position: Position,
-}
-#[derive(Debug, Deserialize)]
-struct LocationEntry<'a> {
-    name: &'a str,
-    zone: &'a str,
-    kind: &'a str,
-    variant: &'a str,
-    uber_group_name: &'a str,
-    uber_group: &'a str,
-    uber_id_name: &'a str,
-    uber_id: &'a str,
-    x: f32,
-    y: f32,
-}
-
-pub fn parse_locations(input: &str) -> Result<Vec<Location>, String> {
-    let mut reader = csv::ReaderBuilder::new()
-        .has_headers(false)
-        .trim(csv::Trim::All)
-        .from_reader(input.as_bytes());
-
-    let mut locations = Vec::with_capacity(389);
-
-    let mut record = csv::StringRecord::new();
-    while reader.read_record(&mut record).map_err(|err| err.to_string())? {
-        let record: LocationEntry = record.deserialize(None).map_err(|err| err.to_string())?;
-
-        let uber_state = UberState::from_parts(record.uber_group, record.uber_id)?;
-        let x = R32::try_from(record.x).map_err(|err| format!("Invalid coordinate {}: {}", record.x, err))?;
-        let y = R32::try_from(record.y).map_err(|err| format!("Invalid coordinate {}: {}", record.y, err))?;
-        let position = Position { x, y };
-        let location = Location { name: record.name.to_owned(), zone: record.zone.to_owned(), uber_state, position };
-
-        locations.push(location);
-    }
-
-    Ok(locations)
-}
-
-#[derive(Debug)]
-pub struct NamedState {
-    pub name: String,
-    pub uber_state: UberState,
-}
-#[derive(Debug, Deserialize)]
-struct StateEntry {
-    name: String,
-    uber_group: String,
-    uber_id: String,
-}
-
-pub fn parse_states(input: &str) -> Result<Vec<NamedState>, String> {
-    let mut reader = csv::ReaderBuilder::new()
-        .has_headers(false)
-        .trim(csv::Trim::All)
-        .from_reader(input.as_bytes());
-
-    let mut states = Vec::with_capacity(97);
-
-    for result in reader.deserialize() {
-        let record: StateEntry = result.map_err(|err| err.to_string())?;
-
-        let uber_state = UberState::from_parts(&record.uber_group, &record.uber_id)?;
-        let state = NamedState { name: record.name, uber_state };
-
-        states.push(state);
-    }
-
-    Ok(states)
 }
