+++ src/world.rs
@@ -148,6 +148,8 @@ impl<'a> World<'a> {
 
 #[cfg(test)]
 mod tests {
+    use crate::languages::logic;
+
     use super::*;
     use super::super::*;
     use world::pool::Pool;
@@ -160,7 +162,7 @@ mod tests {
         let mut settings = Settings::default();
         settings.difficulty = Difficulty::Gorlek;
 
-        let graph = &lexer::parse_logic("areas.wotw", "loc_data.csv", "state_data.csv", &settings, false).unwrap();
+        let graph = &languages::parse_logic("areas.wotw", "loc_data.csv", "state_data.csv", &settings, false).unwrap();
         let mut world = World::new(graph);
         world.player.inventory = Pool::preset().inventory;
         world.player.inventory.grant(Item::SpiritLight(1), 10000);
@@ -176,7 +178,7 @@ mod tests {
             .cloned().collect();
 
         let input = util::read_file("loc_data.csv", "logic").unwrap();
-        let locations = lexer::parser::parse_locations(&input).unwrap();
+        let locations = logic::parse_locations(&input).unwrap();
         let locations: FxHashSet<_> = locations.iter().map(|location| &location.uber_state).cloned().collect();
 
         if !(reached == locations) {
@@ -189,7 +191,7 @@ mod tests {
         let mut settings = Settings::default();
         settings.difficulty = Difficulty::Gorlek;
 
-        let graph = &lexer::parse_logic("areas.wotw", "loc_data.csv", "state_data.csv", &settings, false).unwrap();
+        let graph = &languages::parse_logic("areas.wotw", "loc_data.csv", "state_data.csv", &settings, false).unwrap();
         let mut world = World::new(graph);
 
         world.player.difficulty = Difficulty::Unsafe;
